moved restriciton elements to directory for organization
