+++ Source/Applications/SystemCenter/wwwroot/Scripts/TSX/SystemCenter/CommonComponents/Restrictions/RestrictionTooltip.tsx
